+++ src/client.js
@@ -3,6 +3,7 @@ import DeviceProvider from './components/DeviceProvider'
 import MobileDetect from 'mobile-detect'
 import React from 'react'
 import ReactDOM from 'react-dom'
+import ReactGA from 'react-ga'
 import configureStore from './store/configureStore'
 import createRoutes from './routes'
 import { Provider } from 'react-redux'
@@ -32,6 +33,21 @@ const device = store.getState().device
 
 const routes = createRoutes(history)
 
+if (typeof window !== 'undefined') {
+  // add Google Analytics
+  ReactGA.initialize('UA-69336956-1')
+  ReactGA.set({ page: window.location.pathname })
+}
+
+
+function scrollAndFireTracking() {
+  if(window) {
+    window.scrollTo(0, 0)
+    // send Google Analytics Pageview event on router changed
+    ReactGA.pageview(window.location.pathname)
+  }
+}
+
 // calling `match` is simply for side effects of
 // loading route/component code for the initial location
 // https://github.com/ReactTraining/react-router/blob/v3/docs/guides/ServerRendering.md#async-routes
@@ -39,7 +55,7 @@ match({ history, routes }, (error, redirectLocation, renderProps) => {
   ReactDOM.render((
     <Provider store={store}>
       <DeviceProvider device={device}>
-        <Router {...renderProps} />
+        <Router {...renderProps} onUpdate={scrollAndFireTracking}/>
       </DeviceProvider>
     </Provider>
   ), document.getElementById('root'))
@@ -53,7 +69,7 @@ if (__DEVTOOLS__ && !window.__REDUX_DEVTOOLS_EXTENSION__) {
       <Provider store={store}>
         <DeviceProvider device={device}>
           <div>
-            <Router {...renderProps} />
+            <Router {...renderProps}/>
             <DevTools />
           </div>
         </DeviceProvider>